pyproject.toml
[project]
name = "openclaw"
version = "0.1.0"
description = "AI-powered digital design agency operated by Claude agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "redis[hiredis]>=5.2.0",
    "anthropic>=0.42.0",
    "google-genai>=1.12.0",
    "pywa>=3.9.0",
    "firecrawl-py>=1.0.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.160.0",
    "playwright>=1.49.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "pillow>=11.0.0",
    "structlog>=24.4.0",
    "apscheduler>=3.10.0",
    "python-slugify>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
